OGM-1310 rename parser and test
