+++ mongodb/src/main/java/org/hibernate/ogm/datastore/mongodb/MongoDBDialect.java
@@ -49,7 +49,7 @@ import org.hibernate.ogm.datastore.mongodb.options.impl.AssociationDocumentStora
 import org.hibernate.ogm.datastore.mongodb.options.impl.ReadPreferenceOption;
 import org.hibernate.ogm.datastore.mongodb.options.impl.WriteConcernOption;
 import org.hibernate.ogm.datastore.mongodb.query.impl.MongoDBQueryDescriptor;
-import org.hibernate.ogm.datastore.mongodb.query.parsing.nativequery.impl.StrictNativeQueryParser;
+import org.hibernate.ogm.datastore.mongodb.query.parsing.nativequery.impl.NativeQueryParser;
 import org.hibernate.ogm.datastore.mongodb.type.impl.BinaryAsBsonBinaryGridType;
 import org.hibernate.ogm.datastore.mongodb.type.impl.ObjectIdGridType;
 import org.hibernate.ogm.datastore.mongodb.type.impl.SerializableAsBinaryGridType;
@@ -895,7 +895,7 @@ public class MongoDBDialect extends BaseGridDialect implements QueryableGridDial
 
 	@Override
 	public MongoDBQueryDescriptor parseNativeQuery(String nativeQuery) {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
+		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
 		ParsingResult<MongoDBQueryDescriptor> parseResult = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( nativeQuery );
 		if ( parseResult.hasErrors() ) {
