+++ mongodb/src/test/java/org/hibernate/ogm/datastore/mongodb/test/query/parsing/nativequery/NativeQueryParserTest.java
@@ -6,24 +6,20 @@
  */
 package org.hibernate.ogm.datastore.mongodb.test.query.parsing.nativequery;
 
+import static org.fest.assertions.Assertions.assertThat;
+
+import org.bson.Document;
 import org.hibernate.ogm.datastore.mongodb.query.impl.MongoDBQueryDescriptor;
 import org.hibernate.ogm.datastore.mongodb.query.impl.MongoDBQueryDescriptor.Operation;
-import org.hibernate.ogm.datastore.mongodb.query.parsing.nativequery.impl.MongoDBQueryDescriptorBuilder;
 import org.hibernate.ogm.datastore.mongodb.query.parsing.nativequery.impl.NativeQueryParser;
 import org.hibernate.ogm.datastore.mongodb.utils.DocumentUtil;
 import org.hibernate.ogm.utils.TestForIssue;
-
 import org.junit.Test;
-
-import org.bson.Document;
 import org.parboiled.Parboiled;
+import org.parboiled.parserunners.BasicParseRunner;
 import org.parboiled.parserunners.RecoveringParseRunner;
 import org.parboiled.parserunners.ReportingParseRunner;
-import org.parboiled.support.ParseTreeUtils;
 import org.parboiled.support.ParsingResult;
-
-import static org.fest.assertions.Assertions.assertThat;
-
 
 /**
  * Unit test for {@link NativeQueryParser}.
@@ -39,18 +35,17 @@ public class NativeQueryParserTest {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
 		String match = "{ '$match': {'author' : 'Oscar Wilde' } }";
 		String sort = "{ '$sort': {'name' : -1 } }";
-		ParsingResult<MongoDBQueryDescriptorBuilder> run = new ReportingParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new ReportingParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.aggregate([" + match + ", " + sort + " ])" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.AGGREGATE_PIPELINE );
 		assertThat( queryDescriptor.getPipeline() )
-			.containsExactly(
-					Document.parse( match ),
-					Document.parse( sort )
-					);
+				.containsExactly(
+						Document.parse( match ),
+						Document.parse( sort ) );
 	}
 
 	@Test
@@ -61,37 +56,25 @@ public class NativeQueryParserTest {
 		String unwind = "{'$unwind': '$clicks'}";
 		String group = "{ '$group': {'_id' : '$_id' ,'clicks' : {'$push':'$clicks'} ,'token' : { '$push': '$TOKEN' } } }";
 		String sort = "{ '$sort': { '_id' : -1 } }";
-		String nativeQuery = "db.UserFactualContent.aggregate(["
-						+ match
-						+ "," + unwind
-						+ "," + group
-						+ "," + sort
-						+ "])";
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new ReportingParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() ).run( nativeQuery );
+		String nativeQuery = "db.UserFactualContent.aggregate([" + match + "," + unwind + "," + group + "," + sort + "])";
+		ParsingResult<MongoDBQueryDescriptor> run = new ReportingParseRunner<MongoDBQueryDescriptor>( parser.Query() ).run( nativeQuery );
 
-		System.out.println( ParseTreeUtils.printNodeTree( run ) );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "UserFactualContent" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.AGGREGATE_PIPELINE );
-		assertThat( queryDescriptor.getPipeline() )
-			.containsExactly(
-					Document.parse( match )
-					, Document.parse( unwind )
-					, Document.parse( group )
-					, Document.parse( sort )
-					);
+		assertThat( queryDescriptor.getPipeline() ).containsExactly( Document.parse( match ), Document.parse( unwind ), Document.parse( group ),
+				Document.parse( sort ) );
 	}
 
 	@Test
 	@TestForIssue(jiraKey = "OGM-1247")
 	public void shouldParseDistinctQueryWithFieldOnly() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run = new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
-				.run( "db.Order.distinct('item')" );
+		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() ).run( "db.Order.distinct('item')" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
+
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DISTINCT );
 		assertThat( queryDescriptor.getCriteria() ).isNull();
@@ -105,13 +88,14 @@ public class NativeQueryParserTest {
 	@TestForIssue(jiraKey = "OGM-1247")
 	public void shouldParseDistinctQueryWithCriteriaAndCollation() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run = new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.distinct('item',{'orderId': { '$in': ['XYZ123', '123']}},{'collation' : {'locale' : 'fr'}})" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DISTINCT );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'orderId' : { '$in': ['XYZ123', '123'] } }" ) );
+		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
+				"{ 'orderId' : { '$in': ['XYZ123', '123'] } }" ) );
 		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse( "{'collation' : {'locale' : 'fr'}}" ) );
 		assertThat( queryDescriptor.getProjection() ).isNull();
 		assertThat( queryDescriptor.getOrderBy() ).isNull();
@@ -122,10 +106,10 @@ public class NativeQueryParserTest {
 	@TestForIssue(jiraKey = "OGM-1247")
 	public void shouldParseDistinctQueryWithCollationOnly() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.distinct('item', {}, {'collation' : {'locale' : 'fr'}})" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DISTINCT );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{}" ) );
@@ -142,10 +126,10 @@ public class NativeQueryParserTest {
 		String mapFunction = "function() { emit(this.cust_id, this.price); }";
 		String reduceFuntion = "function(keyCustId, valuesPrices) { return Array.sum(valuesPrices); }";
 		String query = "db.Order.mapReduce('" + mapFunction + "','" + reduceFuntion + "')";
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( query );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.MAP_REDUCE );
 		assertThat( queryDescriptor.getMapFunction() ).isEqualTo( mapFunction );
@@ -160,7 +144,7 @@ public class NativeQueryParserTest {
 	@Test
 	@TestForIssue(jiraKey = "OGM-1246")
 	public void shouldParseComplexMapReduceQuery() {
-		//https://docs.mongodb.com/manual/tutorial/map-reduce-examples/
+		// https://docs.mongodb.com/manual/tutorial/map-reduce-examples/
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
 		String mapFunction = "function() { for (var idx = 0; idx < this.items.length; idx++) { var key = this.items[idx].sku; var value = { count: 1, qty: this.items[idx].qty }; emit(key, value); } }";
 		String reduceFuntion = "function(keySKU, countObjVals) { reducedVal = { count: 0, qty: 0 }; for (var idx = 0; idx < countObjVals.length; idx++) { reducedVal.count += countObjVals[idx].count; reducedVal.qty += countObjVals[idx].qty; } return reducedVal; }";
@@ -177,10 +161,10 @@ public class NativeQueryParserTest {
 				"			'collation' : {'locale' : 'fr'} }";
 
 		String query = "db.Order.mapReduce('" + mapFunction + "','" + reduceFuntion + "'," + options + ")";
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( query );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.MAP_REDUCE );
 		assertThat( queryDescriptor.getMapFunction() ).isEqualTo( mapFunction );
@@ -195,10 +179,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseSimplifiedFindQuery() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "{ \"foo\" : true }" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isNull();
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
@@ -208,12 +192,27 @@ public class NativeQueryParserTest {
 	}
 
 	@Test
+	public void shouldParseObjectsWithNoQuotesInKeys() {
+		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
+				.run( "{ $and: [ { name : 'Portia' }, { author : 'Oscar Wilde' } ] }" );
+
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
+
+		assertThat( queryDescriptor.getCollectionName() ).isNull();
+		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
+		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ $and: [ { name : 'Portia' }, { author : 'Oscar Wilde' } ] }" ) );
+		assertThat( queryDescriptor.getProjection() ).isNull();
+		assertThat( queryDescriptor.getOrderBy() ).isNull();
+	}
+
+	@Test
 	public void shouldParseSimpleQuery() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.find({\"foo\":true})" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
@@ -225,10 +224,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseSimpleQueryUsingSingleQuotes() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.find( { 'foo' : true } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
@@ -240,20 +239,20 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryWithEmptyFind() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.find({})" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( new Document() );
 	}
 
 	@Test
 	public void shouldParseQueryInsertSingleDocument() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.insertOne( { 'item': 'card', 'qty': 15 } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.INSERTONE );
 		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse( "{ 'item': 'card', 'qty': 15 }" ) );
@@ -262,10 +261,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryInsertSingleDocumentAndOptions() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.insertOne( { 'item': 'card', 'qty': 15 }, { 'ordered': true } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.INSERTONE );
 		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse( "{ 'item': 'card', 'qty': 15 }" ) );
@@ -276,35 +275,36 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryInsertMultipleDocuments() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.insertMany( [ { '_id': 11, 'item': 'pencil', 'qty': 50, 'type': 'no.2' }, { 'item': 'pen', 'qty': 20 }, { 'item': 'eraser', 'qty': 25 } ] )" );
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.INSERTMANY );
 		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isNull();
-		assertThat( queryDescriptor.getUpdateOrInsertMany() ).isEqualTo( DocumentUtil.fromJsonArray( "[ { '_id': 11, 'item': 'pencil', 'qty': 50, 'type': 'no.2' }, { 'item': 'pen', 'qty': 20 }, { 'item': 'eraser', 'qty': 25 } ]" ) );
+		assertThat( queryDescriptor.getUpdateOrInsertMany() ).isEqualTo( DocumentUtil.fromJsonArray(
+				"[ { '_id': 11, 'item': 'pencil', 'qty': 50, 'type': 'no.2' }, { 'item': 'pen', 'qty': 20 }, { 'item': 'eraser', 'qty': 25 } ]" ) );
 	}
 
 	@Test
 	public void shouldParseQueryInsertUnkwounDocuments() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.insert( [ { '_id': 11, 'item': 'pencil', 'qty': 50, 'type': 'no.2' }, { 'item': 'pen', 'qty': 20 }, { 'item': 'eraser', 'qty': 25 } ] )" );
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.INSERT );
 		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isNull();
 		assertThat( queryDescriptor.getUpdateOrInsertMany() ).isEqualTo( DocumentUtil.fromJsonArray(
-					"[ { '_id': 11, 'item': 'pencil', 'qty': 50, 'type': 'no.2' }, { 'item': 'pen', 'qty': 20 }, { 'item': 'eraser', 'qty': 25 } ]" ) );
+				"[ { '_id': 11, 'item': 'pencil', 'qty': 50, 'type': 'no.2' }, { 'item': 'pen', 'qty': 20 }, { 'item': 'eraser', 'qty': 25 } ]" ) );
 	}
 
 	@Test
 	public void shouldParseQueryInsertMultipleDocumentsAndOptions() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.insertMany( [ { '_id': 11, 'item': 'pencil', 'qty': 50, 'type': 'no.2' }, { 'item': 'pen', 'qty': 20 }, { 'item': 'eraser', 'qty': 25 } ], { 'ordered': true } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.INSERTMANY );
 		assertThat( queryDescriptor.getUpdateOrInsertMany() ).isEqualTo( DocumentUtil.fromJsonArray(
@@ -316,11 +316,11 @@ public class NativeQueryParserTest {
 	@TestForIssue(jiraKey = "OGM-1313")
 	public void shouldParseQueryDeleteOne() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run = new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>(
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>(
 				parser.Query() )
-				.run( "db.Order.deleteOne( { } )" );
+						.run( "db.Order.deleteOne( { } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DELETEONE );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( new Document() );
@@ -330,11 +330,10 @@ public class NativeQueryParserTest {
 	@TestForIssue(jiraKey = "OGM-1313")
 	public void shouldParseQueryDeleteOneWithFilter() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run = new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>(
-				parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.deleteOne( { 'item': 'card', 'qty': 15 } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DELETEONE );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'item': 'card', 'qty': 15 }" ) );
@@ -345,11 +344,10 @@ public class NativeQueryParserTest {
 	@TestForIssue(jiraKey = "OGM-1313")
 	public void shouldParseQueryDeleteOneWithFilterAndOptions() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run = new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>(
-				parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.deleteOne( { 'item': 'card', 'qty': 15 }, { 'w': 'majority', 'wtimeout' : 100 } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DELETEONE );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'item': 'card', 'qty': 15 }" ) );
@@ -360,11 +358,10 @@ public class NativeQueryParserTest {
 	@TestForIssue(jiraKey = "OGM-1314")
 	public void shouldParseQueryDeleteMany() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run = new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>(
-				parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.deleteMany( { } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DELETEMANY );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( new Document() );
@@ -374,11 +371,10 @@ public class NativeQueryParserTest {
 	@TestForIssue(jiraKey = "OGM-1314")
 	public void shouldParseQueryDeleteManyWithFilter() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run = new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>(
-				parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.deleteMany( { 'item': 'card', 'qty': 15 } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DELETEMANY );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'item': 'card', 'qty': 15 }" ) );
@@ -389,11 +385,10 @@ public class NativeQueryParserTest {
 	@TestForIssue(jiraKey = "OGM-1314")
 	public void shouldParseQueryDeleteManyWithFilterAndOptions() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run = new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>(
-				parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.deleteMany( { 'item': 'card', 'qty': 15 }, { 'w': 'majority', 'wtimeout' : 100 } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DELETEMANY );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'item': 'card', 'qty': 15 }" ) );
@@ -403,10 +398,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryWithEmptyRemove() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.remove( 	{\n 	}\n 	)" ); // Include superfluous whitespace.
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.REMOVE );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( new Document() );
@@ -415,10 +410,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryWithEmptyRemoveAndOptionalJustOne() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.remove({},true)" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.REMOVE );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( new Document() );
@@ -428,10 +423,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryWithEmptyRemoveAndOptions() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.remove( { }, { 'justOne': true } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.REMOVE );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( new Document() );
@@ -441,10 +436,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryUpdate() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.update( { 'name': 'Andy' }, { 'rating': 1, 'score': 1 } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.UPDATE );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'name': 'Andy' }" ) );
@@ -455,10 +450,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryUpdateWithOptions() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.update( { 'name': 'Andy' }, { 'rating': 1, 'score': 1 }, { 'upsert': true } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.UPDATE );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'name': 'Andy' }" ) );
@@ -470,30 +465,29 @@ public class NativeQueryParserTest {
 	@TestForIssue(jiraKey = "OGM-1317")
 	public void shouldParseQueryReplaceOneWithCriteriaWithReplacementWithOptions() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run = new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>(
-				parser.Query() ).run(
-						"db.Order.replaceOne( { 'name' : 'Andy' }, { 'name' : 'Lulu', 'age' : 18, 'score' : 22 }, " +
-								"{ 'upsert': true, 'writeConcern': { 'w': 'majority', 'wtimeout' : 100 }, 'collation': { 'locale': 'fr_CA' } } )" );
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() ).run(
+				"db.Order.replaceOne( { 'name' : 'Andy' }, { 'name' : 'Lulu', 'age' : 18, 'score' : 22 }, " +
+						"{ 'upsert': true, 'writeConcern': { 'w': 'majority', 'wtimeout' : 100 }, 'collation': { 'locale': 'fr_CA' } } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.REPLACEONE );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'name' : 'Andy' }" ) );
-		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse( "{ 'name' : 'Lulu', 'age' : 18, 'score' : 22 }" ) );
-		assertThat( queryDescriptor.getOptions() )
-				.isEqualTo( Document.parse( "{ 'upsert': true, 'writeConcern': { 'w': 'majority', 'wtimeout' : 100 }, 'collation': { 'locale': 'fr_CA' } }" ) );
+		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse(
+				"{ 'name' : 'Lulu', 'age' : 18, 'score' : 22 }" ) );
+		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse(
+				"{ 'upsert': true, 'writeConcern': { 'w': 'majority', 'wtimeout' : 100 }, 'collation': { 'locale': 'fr_CA' } }" ) );
 	}
 
 	@Test
-	@TestForIssue( jiraKey = "OGM-1315" )
+	@TestForIssue(jiraKey = "OGM-1315")
 	public void shouldParseQueryUpdateOne() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run = new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>(
-				parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.updateOne( { 'name' : 'Andy' }, { '$set': { 'score' : 3 } } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.UPDATEONE );
@@ -503,14 +497,13 @@ public class NativeQueryParserTest {
 	}
 
 	@Test
-	@TestForIssue( jiraKey = "OGM-1315" )
+	@TestForIssue(jiraKey = "OGM-1315")
 	public void shouldParseQueryUpdateOneWithOptions() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run = new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>(
-				parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.updateOne( { 'name' : 'Andy' }, { '$set': { 'score' : 3 } }, { 'upsert': true, 'collation': {} } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.UPDATEONE );
@@ -520,13 +513,13 @@ public class NativeQueryParserTest {
 	}
 
 	@Test
-	@TestForIssue( jiraKey = "OGM-1316" )
+	@TestForIssue(jiraKey = "OGM-1316")
 	public void shouldParseQueryUpdateMany() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.updateMany( { 'name': 'Andy' }, { '$mul': { 'score': 5 } } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.UPDATEMANY );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'name': 'Andy' }" ) );
@@ -535,28 +528,29 @@ public class NativeQueryParserTest {
 	}
 
 	@Test
-	@TestForIssue( jiraKey = "OGM-1316" )
+	@TestForIssue(jiraKey = "OGM-1316")
 	public void shouldParseQueryUpdateManyWithOptions() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.updateMany( { 'name' : 'Andy' }, { '$mul': { 'score': 5 } }, { 'upsert': true, 'writeConcern': {'w': 'majority', 'wtimeout' : 100 } } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.UPDATEMANY );
 		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'name' : 'Andy' }" ) );
 		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse( "{ '$mul': { 'score': 5 } }" ) );
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse( "{ 'upsert': true, 'writeConcern': {'w': 'majority', 'wtimeout' : 100 } }" ) );
+		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse(
+				"{ 'upsert': true, 'writeConcern': {'w': 'majority', 'wtimeout' : 100 } }" ) );
 	}
 
 	@Test
 	public void shouldParseQueryFindAndModify() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.findAndModify( { 'query': { 'name': 'Andy' }, 'sort': { 'rating': 1 }, 'update': { '$inc': { 'score': 1 } }, 'upsert': true } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FINDANDMODIFY );
@@ -569,10 +563,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryFindOneWithoutCriteriaNorProjection() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.findOne(  )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FINDONE );
@@ -584,10 +578,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryFindOneWithoutProjection() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.findOne( { \"foo\" : true } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FINDONE );
@@ -599,10 +593,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryFindOneWithCriteriaAndProjection() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.findOne( { \"foo\" : true }, { \"foo\" : 1 } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FINDONE );
@@ -614,10 +608,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryWithProjection() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.find( { \"foo\" : true }, { \"foo\" : 1 } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
@@ -629,10 +623,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryWithWhitespace() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "  db  .  Order  .  find  (  {  \"  foo  \"  :  true  }  ,  {  \"foo\"  :  1  }  )  " );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 		assertThat( run.hasErrors() ).isFalse();
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
@@ -644,12 +638,11 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseQueryWithSeveralConditions() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ReportingParseRunner<MongoDBQueryDescriptorBuilder> runner = new ReportingParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  runner
-				.run( "db.Order.find( { \"foo\" : true, \"bar\" : 42, \"baz\" : \"qux\" } )" );
+		ReportingParseRunner<MongoDBQueryDescriptor> runner = new ReportingParseRunner<MongoDBQueryDescriptor>( parser.Query() );
+		ParsingResult<MongoDBQueryDescriptor> run = runner.run( "db.Order.find( { \"foo\" : true, \"bar\" : 42, \"baz\" : \"qux\" } )" );
 
 		assertThat( run.hasErrors() ).isFalse();
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
@@ -661,10 +654,9 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseCountQuery() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
-				.run( "db.Order.count()" );
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() ).run( "db.Order.count()" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
@@ -676,10 +668,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseCountQueryWithCriteria() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.count( { 'foo' : true } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
@@ -691,14 +683,15 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseCountQueryWithLogicalOperatorOR() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.count( { '$or': [ { 'foo' : true }, { 'bar' : '42' } ] } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ '$or': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
+		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
+				"{ '$or': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
 		assertThat( queryDescriptor.getProjection() ).isNull();
 		assertThat( queryDescriptor.getOrderBy() ).isNull();
 	}
@@ -706,14 +699,15 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseCountQueryWithLogicalOperatorAND() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.count( { '$and': [ { 'foo' : true }, { 'bar' : '42' } ] } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ '$and': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
+		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
+				"{ '$and': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
 		assertThat( queryDescriptor.getProjection() ).isNull();
 		assertThat( queryDescriptor.getOrderBy() ).isNull();
 	}
@@ -721,14 +715,15 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseCountQueryWithLogicalOperatorNOR() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.count( { '$nor': [ { 'foo' : true }, { 'bar' : '42' } ] } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ '$nor': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
+		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
+				"{ '$nor': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
 		assertThat( queryDescriptor.getProjection() ).isNull();
 		assertThat( queryDescriptor.getOrderBy() ).isNull();
 	}
@@ -736,10 +731,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseCountQueryWithLogicalOperatorNOT() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.count( { '$not': { 'foo' : false } } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
@@ -751,14 +746,15 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseFindQueryWithLogicalOperatorOR() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.find( { '$or': [ { 'foo' : true }, { 'bar' : '42' } ] } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ '$or': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
+		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
+				"{ '$or': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
 		assertThat( queryDescriptor.getProjection() ).isNull();
 		assertThat( queryDescriptor.getOrderBy() ).isNull();
 	}
@@ -766,14 +762,15 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldParseFindQueryWithLogicalOperatorAND() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.find( { '$and': [ { 'foo' : true }, { 'bar' : '42' } ] } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ '$and': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
+		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
+				"{ '$and': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
 		assertThat( queryDescriptor.getProjection() ).isNull();
 		assertThat( queryDescriptor.getOrderBy() ).isNull();
 	}
@@ -781,14 +778,15 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldFindCountQueryWithLogicalOperatorNOR() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.find( { '$nor': [ { 'foo' : true }, { 'bar' : '42' } ] } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ '$nor': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
+		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
+				"{ '$nor': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
 		assertThat( queryDescriptor.getProjection() ).isNull();
 		assertThat( queryDescriptor.getOrderBy() ).isNull();
 	}
@@ -796,10 +794,10 @@ public class NativeQueryParserTest {
 	@Test
 	public void shouldFindeCountQueryWithLogicalOperatorNOT() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.Order.find( { '$not': { 'foo' : false } } )" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
@@ -812,10 +810,10 @@ public class NativeQueryParserTest {
 	@TestForIssue(jiraKey = "OGM-900")
 	public void shouldSupportDotInCollectionName() {
 		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptorBuilder> run =  new RecoveringParseRunner<MongoDBQueryDescriptorBuilder>( parser.Query() )
+		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
 				.run( "db.POEM.COM.count()" );
 
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue.build();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
 
 		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "POEM.COM" );
 		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
@@ -823,5 +821,63 @@ public class NativeQueryParserTest {
 		assertThat( queryDescriptor.getOrderBy() ).isNull();
 	}
 
+	@Test
+	public void shouldFailOnUnknownOperation() {
+		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
+		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
+				.run( " db.Order.fidn() " );
 
+		assertThat( run.matched ).isFalse();
+		assertThat( run.hasErrors() ).isTrue();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
+		assertThat( queryDescriptor ).isNull();
+	}
+
+	@Test
+	public void shouldFailOnInvalidCommand() {
+		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
+		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
+				.run( "Order.updateMany( { 'name': 'Andy' }, { '$mul': { 'score': 5 } } )" );
+
+		assertThat( run.matched ).isFalse();
+		assertThat( run.hasErrors() ).isTrue();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
+		assertThat( queryDescriptor ).isNull();
+	}
+
+	@Test
+	public void shouldFailOnMissingCollectionName() {
+		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
+		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
+				.run( "db.find( { 'name': 'Andy' } )" );
+
+		assertThat( run.matched ).isFalse();
+		assertThat( run.hasErrors() ).isTrue();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
+		assertThat( queryDescriptor ).isNull();
+	}
+
+	@Test
+	public void shouldFailOnCollectionInvalidName() {
+		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
+		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
+				.run( "db.findOne.update( { 'name': 'Andy' }, { '$mul': { 'score': 5 } } )" );
+
+		assertThat( run.matched ).isFalse();
+		assertThat( run.hasErrors() ).isTrue();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
+		assertThat( queryDescriptor ).isNull();
+	}
+
+	@Test
+	public void shouldFailOnInvalidFirstIdentifier() {
+		NativeQueryParser parser = Parboiled.createParser( NativeQueryParser.class );
+		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
+				.run( "Order.update( { 'name': 'Andy' }, { '$mul': { 'score': 5 } } )" );
+
+		assertThat( run.matched ).isFalse();
+		assertThat( run.hasErrors() ).isTrue();
+		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
+		assertThat( queryDescriptor ).isNull();
+	}
 }
