--- mongodb/src/test/java/org/hibernate/ogm/datastore/mongodb/test/query/parsing/nativequery/StrictNativeQueryParserTest.java
@@ -1,868 +1,0 @@
-/*
- * Hibernate OGM, Domain model persistence for NoSQL datastores
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.ogm.datastore.mongodb.test.query.parsing.nativequery;
-
-import static org.fest.assertions.Assertions.assertThat;
-
-import org.bson.Document;
-import org.hibernate.ogm.datastore.mongodb.query.impl.MongoDBQueryDescriptor;
-import org.hibernate.ogm.datastore.mongodb.query.impl.MongoDBQueryDescriptor.Operation;
-import org.hibernate.ogm.datastore.mongodb.query.parsing.nativequery.impl.StrictNativeQueryParser;
-import org.hibernate.ogm.datastore.mongodb.utils.DocumentUtil;
-import org.hibernate.ogm.utils.TestForIssue;
-import org.junit.Test;
-import org.parboiled.Parboiled;
-import org.parboiled.parserunners.BasicParseRunner;
-import org.parboiled.parserunners.RecoveringParseRunner;
-import org.parboiled.parserunners.ReportingParseRunner;
-import org.parboiled.support.ParsingResult;
-
-/**
- * Unit test for {@link StrictNativeQueryParser}.
- *
- * @author Gunnar Morling
- * @see <a href="https://docs.mongodb.com/manual/tutorial/insert-documents/"> InsertOne documents in 3.4 API</a>
- */
-public class StrictNativeQueryParserTest {
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1024")
-	public void shouldParseSimplifiedAggregateQuery() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		String match = "{ '$match': {'author' : 'Oscar Wilde' } }";
-		String sort = "{ '$sort': {'name' : -1 } }";
-		ParsingResult<MongoDBQueryDescriptor> run = new ReportingParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.aggregate([" + match + ", " + sort + " ])" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.AGGREGATE_PIPELINE );
-		assertThat( queryDescriptor.getPipeline() )
-				.containsExactly(
-						Document.parse( match ),
-						Document.parse( sort ) );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1024")
-	public void shouldParseComplexAggregateQuery() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		String match = "{ '$match': { 'NAME':{'$regex':'Bangalore', '$options': 'i'}}}";
-		String unwind = "{'$unwind': '$clicks'}";
-		String group = "{ '$group': {'_id' : '$_id' ,'clicks' : {'$push':'$clicks'} ,'token' : { '$push': '$TOKEN' } } }";
-		String sort = "{ '$sort': { '_id' : -1 } }";
-		String nativeQuery = "db.UserFactualContent.aggregate([" + match + "," + unwind + "," + group + "," + sort + "])";
-		ParsingResult<MongoDBQueryDescriptor> run = new ReportingParseRunner<MongoDBQueryDescriptor>( parser.Query() ).run( nativeQuery );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "UserFactualContent" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.AGGREGATE_PIPELINE );
-		assertThat( queryDescriptor.getPipeline() ).containsExactly( Document.parse( match ), Document.parse( unwind ), Document.parse( group ),
-				Document.parse( sort ) );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1247")
-	public void shouldParseDistinctQueryWithFieldOnly() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() ).run( "db.Order.distinct('item')" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DISTINCT );
-		assertThat( queryDescriptor.getCriteria() ).isNull();
-		assertThat( queryDescriptor.getOptions() ).isNull();
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-		assertThat( queryDescriptor.getDistinctFieldName() ).isEqualTo( "item" );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1247")
-	public void shouldParseDistinctQueryWithCriteriaAndCollation() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.distinct('item',{'orderId': { '$in': ['XYZ123', '123']}},{'collation' : {'locale' : 'fr'}})" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DISTINCT );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
-				"{ 'orderId' : { '$in': ['XYZ123', '123'] } }" ) );
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse( "{'collation' : {'locale' : 'fr'}}" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-		assertThat( queryDescriptor.getDistinctFieldName() ).isEqualTo( "item" );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1247")
-	public void shouldParseDistinctQueryWithCollationOnly() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.distinct('item', {}, {'collation' : {'locale' : 'fr'}})" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DISTINCT );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{}" ) );
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse( "{'collation' : {'locale' : 'fr'} }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-		assertThat( queryDescriptor.getDistinctFieldName() ).isEqualTo( "item" );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1246")
-	public void shouldParseSimpleMapReduceQuery() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		String mapFunction = "function() { emit(this.cust_id, this.price); }";
-		String reduceFuntion = "function(keyCustId, valuesPrices) { return Array.sum(valuesPrices); }";
-		String query = "db.Order.mapReduce('" + mapFunction + "','" + reduceFuntion + "')";
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( query );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.MAP_REDUCE );
-		assertThat( queryDescriptor.getMapFunction() ).isEqualTo( mapFunction );
-		assertThat( queryDescriptor.getReduceFunction() ).isEqualTo( reduceFuntion );
-		assertThat( queryDescriptor.getOptions() ).isNull();
-		assertThat( queryDescriptor.getCriteria() ).isNull();
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-		assertThat( queryDescriptor.getDistinctFieldName() ).isNull();
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1246")
-	public void shouldParseComplexMapReduceQuery() {
-		// https://docs.mongodb.com/manual/tutorial/map-reduce-examples/
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		String mapFunction = "function() { for (var idx = 0; idx < this.items.length; idx++) { var key = this.items[idx].sku; var value = { count: 1, qty: this.items[idx].qty }; emit(key, value); } }";
-		String reduceFuntion = "function(keySKU, countObjVals) { reducedVal = { count: 0, qty: 0 }; for (var idx = 0; idx < countObjVals.length; idx++) { reducedVal.count += countObjVals[idx].count; reducedVal.qty += countObjVals[idx].qty; } return reducedVal; }";
-		String finalizeFunction = "function (key, reducedVal) { reducedVal.avg = reducedVal.qty/reducedVal.count; return reducedVal; }";
-		String options = "{ 'out': 'mycollection'," +
-				"			'query': {}," +
-				"			'sort' : {}," +
-				"			'limit' : 1000," +
-				"			'finalize' : '" + finalizeFunction + "'," +
-				"			'scope' : {}," +
-				"			'jsMode' : true," +
-				"			'verbose' : false," +
-				"			'bypassDocumentValidation' : false," +
-				"			'collation' : {'locale' : 'fr'} }";
-
-		String query = "db.Order.mapReduce('" + mapFunction + "','" + reduceFuntion + "'," + options + ")";
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( query );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.MAP_REDUCE );
-		assertThat( queryDescriptor.getMapFunction() ).isEqualTo( mapFunction );
-		assertThat( queryDescriptor.getReduceFunction() ).isEqualTo( reduceFuntion );
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse( options ) );
-		assertThat( queryDescriptor.getCriteria() ).isNull();
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-		assertThat( queryDescriptor.getDistinctFieldName() ).isNull();
-	}
-
-	@Test
-	public void shouldParseSimplifiedFindQuery() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "{ \"foo\" : true }" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isNull();
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ \"foo\" : true }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseSimpleQuery() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.find({\"foo\":true})" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ \"foo\" : true }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseSimpleQueryUsingSingleQuotes() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.find( { 'foo' : true } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ \"foo\" : true }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseQueryWithEmptyFind() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.find({})" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( new Document() );
-	}
-
-	@Test
-	public void shouldParseQueryInsertSingleDocument() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.insertOne( { 'item': 'card', 'qty': 15 } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.INSERTONE );
-		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse( "{ 'item': 'card', 'qty': 15 }" ) );
-	}
-
-	@Test
-	public void shouldParseQueryInsertSingleDocumentAndOptions() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.insertOne( { 'item': 'card', 'qty': 15 }, { 'ordered': true } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.INSERTONE );
-		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse( "{ 'item': 'card', 'qty': 15 }" ) );
-		assertThat( queryDescriptor.getUpdateOrInsertMany() ).isNull();
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse( "{ ordered: true })" ) );
-	}
-
-	@Test
-	public void shouldParseQueryInsertMultipleDocuments() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.insertMany( [ { '_id': 11, 'item': 'pencil', 'qty': 50, 'type': 'no.2' }, { 'item': 'pen', 'qty': 20 }, { 'item': 'eraser', 'qty': 25 } ] )" );
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.INSERTMANY );
-		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isNull();
-		assertThat( queryDescriptor.getUpdateOrInsertMany() ).isEqualTo( DocumentUtil.fromJsonArray(
-				"[ { '_id': 11, 'item': 'pencil', 'qty': 50, 'type': 'no.2' }, { 'item': 'pen', 'qty': 20 }, { 'item': 'eraser', 'qty': 25 } ]" ) );
-	}
-
-	@Test
-	public void shouldParseQueryInsertUnkwounDocuments() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.insert( [ { '_id': 11, 'item': 'pencil', 'qty': 50, 'type': 'no.2' }, { 'item': 'pen', 'qty': 20 }, { 'item': 'eraser', 'qty': 25 } ] )" );
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.INSERT );
-		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isNull();
-		assertThat( queryDescriptor.getUpdateOrInsertMany() ).isEqualTo( DocumentUtil.fromJsonArray(
-				"[ { '_id': 11, 'item': 'pencil', 'qty': 50, 'type': 'no.2' }, { 'item': 'pen', 'qty': 20 }, { 'item': 'eraser', 'qty': 25 } ]" ) );
-	}
-
-	@Test
-	public void shouldParseQueryInsertMultipleDocumentsAndOptions() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.insertMany( [ { '_id': 11, 'item': 'pencil', 'qty': 50, 'type': 'no.2' }, { 'item': 'pen', 'qty': 20 }, { 'item': 'eraser', 'qty': 25 } ], { 'ordered': true } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.INSERTMANY );
-		assertThat( queryDescriptor.getUpdateOrInsertMany() ).isEqualTo( DocumentUtil.fromJsonArray(
-				"[ { '_id': 11, 'item': 'pencil', 'qty': 50, 'type': 'no.2' }, { 'item': 'pen', 'qty': 20 }, { 'item': 'eraser', 'qty': 25 } ]" ) );
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse( "{ ordered: true })" ) );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1313")
-	public void shouldParseQueryDeleteOne() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>(
-				parser.Query() )
-						.run( "db.Order.deleteOne( { } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DELETEONE );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( new Document() );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1313")
-	public void shouldParseQueryDeleteOneWithFilter() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.deleteOne( { 'item': 'card', 'qty': 15 } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DELETEONE );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'item': 'card', 'qty': 15 }" ) );
-		assertThat( queryDescriptor.getOptions() ).isNull();
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1313")
-	public void shouldParseQueryDeleteOneWithFilterAndOptions() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.deleteOne( { 'item': 'card', 'qty': 15 }, { 'w': 'majority', 'wtimeout' : 100 } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DELETEONE );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'item': 'card', 'qty': 15 }" ) );
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse( "{ 'w': 'majority', 'wtimeout' : 100 }" ) );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1314")
-	public void shouldParseQueryDeleteMany() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.deleteMany( { } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DELETEMANY );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( new Document() );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1314")
-	public void shouldParseQueryDeleteManyWithFilter() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.deleteMany( { 'item': 'card', 'qty': 15 } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DELETEMANY );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'item': 'card', 'qty': 15 }" ) );
-		assertThat( queryDescriptor.getOptions() ).isNull();
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1314")
-	public void shouldParseQueryDeleteManyWithFilterAndOptions() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.deleteMany( { 'item': 'card', 'qty': 15 }, { 'w': 'majority', 'wtimeout' : 100 } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.DELETEMANY );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'item': 'card', 'qty': 15 }" ) );
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse( "{ 'w': 'majority', 'wtimeout' : 100 }" ) );
-	}
-
-	@Test
-	public void shouldParseQueryWithEmptyRemove() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.remove( 	{\n 	}\n 	)" ); // Include superfluous whitespace.
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.REMOVE );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( new Document() );
-	}
-
-	@Test
-	public void shouldParseQueryWithEmptyRemoveAndOptionalJustOne() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.remove({},true)" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.REMOVE );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( new Document() );
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse( "{ \"justOne\" : true }" ) );
-	}
-
-	@Test
-	public void shouldParseQueryWithEmptyRemoveAndOptions() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.remove( { }, { 'justOne': true } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.REMOVE );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( new Document() );
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse( "{ \"justOne\" : true }" ) );
-	}
-
-	@Test
-	public void shouldParseQueryUpdate() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.update( { 'name': 'Andy' }, { 'rating': 1, 'score': 1 } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.UPDATE );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'name': 'Andy' }" ) );
-		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse( "{ 'rating': 1, 'score': 1 }" ) );
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseQueryUpdateWithOptions() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.update( { 'name': 'Andy' }, { 'rating': 1, 'score': 1 }, { 'upsert': true } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.UPDATE );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'name': 'Andy' }" ) );
-		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse( "{ 'rating': 1, 'score': 1 }" ) );
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse( "{ 'upsert': true }" ) );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1317")
-	public void shouldParseQueryReplaceOneWithCriteriaWithReplacementWithOptions() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() ).run(
-				"db.Order.replaceOne( { 'name' : 'Andy' }, { 'name' : 'Lulu', 'age' : 18, 'score' : 22 }, " +
-						"{ 'upsert': true, 'writeConcern': { 'w': 'majority', 'wtimeout' : 100 }, 'collation': { 'locale': 'fr_CA' } } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.REPLACEONE );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'name' : 'Andy' }" ) );
-		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse(
-				"{ 'name' : 'Lulu', 'age' : 18, 'score' : 22 }" ) );
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse(
-				"{ 'upsert': true, 'writeConcern': { 'w': 'majority', 'wtimeout' : 100 }, 'collation': { 'locale': 'fr_CA' } }" ) );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1315")
-	public void shouldParseQueryUpdateOne() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.updateOne( { 'name' : 'Andy' }, { '$set': { 'score' : 3 } } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.UPDATEONE );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'name' : 'Andy' }" ) );
-		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse( "{ '$set': { 'score' : 3 } }" ) );
-		assertThat( queryDescriptor.getOptions() ).isNull();
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1315")
-	public void shouldParseQueryUpdateOneWithOptions() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.updateOne( { 'name' : 'Andy' }, { '$set': { 'score' : 3 } }, { 'upsert': true, 'collation': {} } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.UPDATEONE );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'name' : 'Andy' }" ) );
-		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse( "{ '$set': { 'score' : 3 } }" ) );
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse( "{ 'upsert': true, 'collation': {} }" ) );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1316")
-	public void shouldParseQueryUpdateMany() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.updateMany( { 'name': 'Andy' }, { '$mul': { 'score': 5 } } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.UPDATEMANY );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'name': 'Andy' }" ) );
-		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse( "{ '$mul': { 'score': 5 } }" ) );
-		assertThat( queryDescriptor.getOptions() ).isNull();
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-1316")
-	public void shouldParseQueryUpdateManyWithOptions() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.updateMany( { 'name' : 'Andy' }, { '$mul': { 'score': 5 } }, { 'upsert': true, 'writeConcern': {'w': 'majority', 'wtimeout' : 100 } } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.UPDATEMANY );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'name' : 'Andy' }" ) );
-		assertThat( queryDescriptor.getUpdateOrInsertOne() ).isEqualTo( Document.parse( "{ '$mul': { 'score': 5 } }" ) );
-		assertThat( queryDescriptor.getOptions() ).isEqualTo( Document.parse(
-				"{ 'upsert': true, 'writeConcern': {'w': 'majority', 'wtimeout' : 100 } }" ) );
-	}
-
-	@Test
-	public void shouldParseQueryFindAndModify() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.findAndModify( { 'query': { 'name': 'Andy' }, 'sort': { 'rating': 1 }, 'update': { '$inc': { 'score': 1 } }, 'upsert': true } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FINDANDMODIFY );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
-				"{ 'query': { 'name': 'Andy' }, 'sort': { 'rating': 1 }, 'update': { '$inc': { 'score': 1 } }, 'upsert': true }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseQueryFindOneWithoutCriteriaNorProjection() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.findOne(  )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FINDONE );
-		assertThat( queryDescriptor.getCriteria() ).isNull();
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseQueryFindOneWithoutProjection() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.findOne( { \"foo\" : true } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FINDONE );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ \"foo\" : true }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseQueryFindOneWithCriteriaAndProjection() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.findOne( { \"foo\" : true }, { \"foo\" : 1 } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FINDONE );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ \"foo\" : true }" ) );
-		assertThat( queryDescriptor.getProjection() ).isEqualTo( Document.parse( "{ \"foo\" : 1 }" ) );
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseQueryWithProjection() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.find( { \"foo\" : true }, { \"foo\" : 1 } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ \"foo\" : true }" ) );
-		assertThat( queryDescriptor.getProjection() ).isEqualTo( Document.parse( "{ \"foo\" : 1 }" ) );
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseQueryWithWhitespace() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "  db  .  Order  .  find  (  {  \"  foo  \"  :  true  }  ,  {  \"foo\"  :  1  }  )  " );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( run.hasErrors() ).isFalse();
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ \"  foo  \" : true }" ) );
-		assertThat( queryDescriptor.getProjection() ).isEqualTo( Document.parse( "{ \"foo\" : 1 }" ) );
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseQueryWithSeveralConditions() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ReportingParseRunner<MongoDBQueryDescriptor> runner = new ReportingParseRunner<MongoDBQueryDescriptor>( parser.Query() );
-		ParsingResult<MongoDBQueryDescriptor> run = runner.run( "db.Order.find( { \"foo\" : true, \"bar\" : 42, \"baz\" : \"qux\" } )" );
-
-		assertThat( run.hasErrors() ).isFalse();
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ \"foo\" : true, \"bar\" : 42, \"baz\" : \"qux\" }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseCountQuery() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() ).run( "db.Order.count()" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
-		assertThat( queryDescriptor.getCriteria() ).isNull();
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseCountQueryWithCriteria() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.count( { 'foo' : true } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ 'foo' : true }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseCountQueryWithLogicalOperatorOR() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.count( { '$or': [ { 'foo' : true }, { 'bar' : '42' } ] } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
-				"{ '$or': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseCountQueryWithLogicalOperatorAND() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.count( { '$and': [ { 'foo' : true }, { 'bar' : '42' } ] } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
-				"{ '$and': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseCountQueryWithLogicalOperatorNOR() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.count( { '$nor': [ { 'foo' : true }, { 'bar' : '42' } ] } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
-				"{ '$nor': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseCountQueryWithLogicalOperatorNOT() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.count( { '$not': { 'foo' : false } } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ '$not': { 'foo' : false } } }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseFindQueryWithLogicalOperatorOR() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.find( { '$or': [ { 'foo' : true }, { 'bar' : '42' } ] } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
-				"{ '$or': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldParseFindQueryWithLogicalOperatorAND() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.find( { '$and': [ { 'foo' : true }, { 'bar' : '42' } ] } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
-				"{ '$and': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldFindCountQueryWithLogicalOperatorNOR() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.find( { '$nor': [ { 'foo' : true }, { 'bar' : '42' } ] } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse(
-				"{ '$nor': [ { 'foo' : true }, { 'bar' : '42' } ] } }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldFindeCountQueryWithLogicalOperatorNOT() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new RecoveringParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.Order.find( { '$not': { 'foo' : false } } )" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "Order" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.FIND );
-		assertThat( queryDescriptor.getCriteria() ).isEqualTo( Document.parse( "{ '$not': { 'foo' : false } } }" ) );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "OGM-900")
-	public void shouldSupportDotInCollectionName() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.POEM.COM.count()" );
-
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-
-		assertThat( queryDescriptor.getCollectionName() ).isEqualTo( "POEM.COM" );
-		assertThat( queryDescriptor.getOperation() ).isEqualTo( Operation.COUNT );
-		assertThat( queryDescriptor.getProjection() ).isNull();
-		assertThat( queryDescriptor.getOrderBy() ).isNull();
-	}
-
-	@Test
-	public void shouldFailOnUnknownOperation() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( " db.Order.fidn() " );
-
-		assertThat( run.matched ).isFalse();
-		assertThat( run.hasErrors() ).isTrue();
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor ).isNull();
-	}
-
-	@Test
-	public void shouldFailOnInvalidCommand() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "Order.updateMany( { 'name': 'Andy' }, { '$mul': { 'score': 5 } } )" );
-
-		assertThat( run.matched ).isFalse();
-		assertThat( run.hasErrors() ).isTrue();
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor ).isNull();
-	}
-
-	@Test
-	public void shouldFailOnMissingCollectionName() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.find( { 'name': 'Andy' } )" );
-
-		assertThat( run.matched ).isFalse();
-		assertThat( run.hasErrors() ).isTrue();
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor ).isNull();
-	}
-
-	@Test
-	public void shouldFailOnCollectionInvalidName() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "db.findOne.update( { 'name': 'Andy' }, { '$mul': { 'score': 5 } } )" );
-
-		assertThat( run.matched ).isFalse();
-		assertThat( run.hasErrors() ).isTrue();
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor ).isNull();
-	}
-
-	@Test
-	public void shouldFailOnInvalidFirstIdentifier() {
-		StrictNativeQueryParser parser = Parboiled.createParser( StrictNativeQueryParser.class );
-		ParsingResult<MongoDBQueryDescriptor> run = new BasicParseRunner<MongoDBQueryDescriptor>( parser.Query() )
-				.run( "Order.update( { 'name': 'Andy' }, { '$mul': { 'score': 5 } } )" );
-
-		assertThat( run.matched ).isFalse();
-		assertThat( run.hasErrors() ).isTrue();
-		MongoDBQueryDescriptor queryDescriptor = run.resultValue;
-		assertThat( queryDescriptor ).isNull();
-	}
-}
